change tg_bot_id on BigInteger(int64)

+++ app/database/models/tables/tg_bot.py
@@ -1,6 +1,6 @@
 from typing import Optional
 
-from sqlalchemy import Column, Integer, String, Text, ForeignKey
+from sqlalchemy import Column, Integer, String, Text, ForeignKey, BigInteger
 
 from database.connection import Base
 
@@ -8,7 +8,7 @@ from database.connection import Base
 class TGBot(Base):
     __tablename__ = 'tg_bots'
 
-    tg_bot_id: int = Column(Integer, primary_key=True, autoincrement=False)
+    tg_bot_id: int = Column(BigInteger, primary_key=True, autoincrement=False)
     tg_token: int = Column(String(46), nullable=False)
     start_message: Optional[str] = Column(Text, nullable=True)
     qiwi_txn: Optional[int] = Column(Integer, nullable=True)
